backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
scipy
librosa
soundfile
imagehash
Pillow
sqlalchemy
aiosqlite
faiss-cpu
torch
torchvision
torchaudio
transformers
python-jose[cryptography]
pydantic
aiofiles
